ColorAlchemy/requirements.txt
flask==1.1.2
torch==1.7.1
fastai==2.3.1
opencv-python==4.5.2.54
Pillow==8.2.0
